report page rerouting
